requirements.txt
# Flask Web Framework
flask==2.3.3
werkzeug==2.3.7

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Groq LLM API
groq==0.4.0

# Optional: Production Server
gunicorn==21.2.0
